+++ evergreen.yaml
@@ -69,11 +69,11 @@ tasks:
 
   - <<: *run-build
     tags: ["test"]
-    name: test-buildsystem-generator
+    name: test-metabuild-generator
 
   - <<: *run-build
     tags: ["test"]
-    name: test-buildsystem-model
+    name: test-metabuild-model
 
   - name: compile-base
     tags: ["legacy-check"]
@@ -84,11 +84,11 @@ tasks:
 
   - <<: *run-build
     tags: ["report"]
-    name: lint-buildsystem-generator
+    name: lint-metabuild-generator
 
   - <<: *run-build
     tags: ["report"]
-    name: lint-buildsystem-model
+    name: lint-metabuild-model
 
   - <<: *run-build
     tags: ["report"]
