+++ metabuild/generator/golang_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/evergreen-ci/shrub"
-	"github.com/mongodb/jasper/buildsystem/model"
+	"github.com/mongodb/jasper/metabuild/model"
 	"github.com/mongodb/jasper/util"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
