+++ metabuild/generator/make.go
@@ -2,7 +2,7 @@ package generator
 
 import (
 	"github.com/evergreen-ci/shrub"
-	"github.com/mongodb/jasper/buildsystem/model"
+	"github.com/mongodb/jasper/metabuild/model"
 	"github.com/pkg/errors"
 )
