+++ metabuild/generator/make_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/evergreen-ci/shrub"
-	"github.com/mongodb/jasper/buildsystem/model"
+	"github.com/mongodb/jasper/metabuild/model"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
